rename cards.js to events.js
