+++ frontend/src/game/DeckSettings.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import App from "../app";
-import {getZone, getZoneDisplayName} from "../cards";
+import {getZone, getZoneDisplayName} from "../events";
 import {BASICS, ZONE_MAIN, ZONE_SIDEBOARD} from "../gamestate";
 import {Select} from "../utils";
