+++ frontend/src/init.js
@@ -1,5 +1,5 @@
 import App from "./app";
-import "./cards";
+import "./events";
 import { default as router } from "./router";
 
 App.init(router);
